requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
requests==2.32.3
pydantic==2.9.2
python-dotenv==1.0.1
langchain==0.3.7
langchain-openai==0.2.6
